Fuel tanks

Liquid hydrocarbon fuel is assumed to be stored in the interior of the wings and no additional tanks are needed. The weight of the fuel is accounted for while sizing the wing structure. See structures.surfw.

However, alternate fuels such as cryogenic liquid hydrogen require additional storage tanks that are insulated pressure vessels.

TASOPT.structures.tanksize — Function

    tanksize(gee, rhoFuel, deltap,
            Rfuse, dRfuse, hconvgas, h_LH2, Tfuel, Tair,
            h_v, t_cond, k, hconvair, time_flight, fstring,ffadd,
            wfb, nfweb, sigskin, rho_insul, rhoskin, Wfuel, threshold_percent, clearance_fuse, AR)

tanksize sizes a cryogenic fuel tank for a LH-fueled aircraft

🔃 Inputs and Outputs

**Inputs:**
- `gee::Float64`: Gravitational acceleration (m/s^2).
- `rhoFuel::Float64`: Density of fuel (kg/m^3).
- `deltap::Float64`: Allowed pressure difference in vessel (Pa).
- `Rfuse::Float64`: Fuselage radius (m).
- `rho_insul::Array{Float64}`: Array of insulation layer densities (kg/m3).
- `dRfuse::Float64`: Accounts for flatness at the bottom of the fuselage (m).
- `xshell1::Float64`: Start x-coordinate of the tank (m).
- `xshell2::Float64`: End x-coordinate of the tank (m).
- `Thermal::Array{Float64}`: Conductivity array k (W/(m*K)) for each MLI layer.
- `hconvgas::Float64`: Convective coefficient of insulating purged gas (e.g., N2) (W/m2*K).
- `hconvair::Float64`: Convective coefficient of ambient air (W/m2*K).
- `Thickness::Array{Float64}`: Thickness array t (m) for each MLI layer.
- `h_LH2::Float64`: LH2 convective coefficient (W/m2*K).
- `Tfuel::Float64`: Fuel temperature (K).
- `Tair::Float64`: Ambient temperature (K).
- `r_tank::Float64`: Tank outer radius (m).
- `h_v::Float64`: Heat of vaporization of liquid hydrogen (J/kg).
- `r_gas::Float64`: Inner radius of the gas-purged chamber (m).
- `threshold_percent::Float64`: Max allowed percentage of fuel that is allowed to boil off.
- `mode::Char`: '1' means optimize m_boiloff, '2' means find m_boiloff based on given thickness.

**Outputs:**
- `Wtank_total::Float64`: Total weight of the tank including fuel (N).
- `thickness_insul::Float64`: Total thickness of the insulation (m).
- `lshell::Float64`: Length of the tank (m).
- `mdot_boiloff::Float64`: Mass boiled off during the mission flight (kg).
- `Vfuel::Float64`: Volume of fuel (m^3).
- `Wfuel_tot::Float64`: Weight of fuel (N).
- `m_boiloff::Float64`: Mass boiled off (kg).
- `tskin::Float64`: Thickness of the tank's skin (m).
- `t_head::Float64`: Thickness of the tank's head (m).
- `Rtank::Float64`: Radius of the tank (m).
- `Whead::Float64`: Weight of the tank's head (N).
- `Wcyl::Float64`: Weight of the tank's cylinder (N).
- `Winsul_sum::Float64`: Sum of the insulation weight (N).
- `Winsul::Float64`: Weight of insulation (N).
- `l_tank::Float64`: Length of the tank (m).
- `Wtank::Float64`: Weight of the tank structure (N).

See here.

TASOPT.structures.tankWmech — Function

  tankWmech(gee, ρfuel,
        ftankstiff, ftankadd, Δp,
        Rfuse, dRfuse, wfb, nfweb,
        sigskin, rho_insul, rhoskin,
        Wfuel, m_boiloff, t_cond, clearance_fuse, AR)

tankWmech calculates the weight of the cryogenic fuel tank for a LH-fueled aircraft.

🔃 Inputs and Outputs

Inputs:

gee::Float64: Gravitational acceleration (m/s^2).

ρfuel::Float64: Fuel density (kg/m^3).

ftankstiff::Float64: Tank stiffness factor.

ftankadd::Float64: Additional mass factor for the tank.

Δp::Float64: Pressure difference (Pa).

Rfuse::Float64: Fuselage radius (m).

dRfuse::Float64: Subtraction factor accounting for fuselage flatness (m).

wfb: Parameters for multiple-bubble configuration.

nfweb: Number of bubbles.

sigskin::Float64: Material property.

rho_insul::Array{Float64}: Array of insulation layer densities (kg/m3).

rhoskin::Float64: Material property.

Wfuel::Float64: Weight of fuel (N).

m_boiloff::Float64: Mass boiled off during the mission flight (kg).

t_cond::Array{Float64,1}: Thickness of insulation layers (m).

clearance_fuse::Float64: Clearance for the fuselage (m).

AR::Float64: Aspect ratio.

Outputs:

Wtank_total::Float64: Total tank weight including fuel (N).

l_cyl::Float64: Length of the cylindrical portion (m).

tskin::Float64: Thickness of the tank's skin (m).

Rtank_outer::Float64: Outer radius of the tank (m).

Vfuel::Float64: Volume of fuel (m^3).

Wtank::Float64: Weight of the empty tank (N).

Wfuel_tot::Float64: Total weight of fuel (N).

Winsul_sum::Float64: Sum of insulation weight (N).

t_head::Float64: Thickness of the tank's head (m).

Whead::Float64: Weight of the tank's head (N).

Wcyl::Float64: Weight of the tank's cylinder (N).

Winsul::Float64: Weight of insulation (N).

Shead_insul::Float64: Insulated surface area of the head (m^2).

l_tank::Float64: Total length of the tank (m).

NOTE: Al alloy 2219 has been recommended as tank material (from H2 tank paper in OneNote)

See here.

TASOPT.structures.tankWthermal — Function

  tankWthermal(l_cyl, r_tank, Shead,
        hconvgas, h_LH2, hconvair, 
        t_cond, k, Tfuel, Tair, 
        h_v:, time_flight)

tankWthermal calculates the boil-off rate of LH2 for a given insulation thickness.

This subroutine does not size the thermal insulation layers but rather calculates the boil-off rate of fLH2, for a given insulation thickness

🔃 Inputs and Outputs

Inputs:

l_cyl::Float64: Length of the tank (m).

r_tank::Float64: Tank outer radius (m).

Shead::Array{Float64,1}: Array of surface areas of each layer of the end/head of the tank [m²].

hconvgas::Float64: Convective coefficient of insulating purged gas (W/m²*K).

h_LH2::Float64: LH2 convective coefficient (W/m²*K).

hconvair::Float64: Convective coefficient of ambient air (W/m²*K).

t_cond::Array{Float64,1}: Array of thickness of each layer in MLI (m).

k::Array{Float64,1}: Thermal conductivity array (W/(m*K)) comprising k values for each MLI layer.

Tfuel::Float64: Fuel temperature (K).

Tair::Float64: Ambient temperature (K).

h_v::Float64: Heat of vaporization of liquid hydrogen (J/kg).

time_flight::Float64: Time of flight (s).

Outputs:

m_boiloff::Float64: Boil-off LH2 mass (kg).

mdot_boiloff::Float64: Boil-off rate of LH2 (kg/s).

See here.