Stability and balance

TASOPT.balance — Function

  balance(pari, parg, para, rfuel, rpay, ξpay, itrim)

Makes one of three (or none) changes to achieve pitch trim calculates resulting CG, CP, NP locations.

Inputs:

pari[.] integer flag array

parg[.] geometry parameter array

para[.] aero parameter array

rfuel fuel fraction Wfuelactual/WfuelMTOW

rpay payload fraction Wpayactual/WpayMTOW

ξpay partial-payload packing location

= 0.0 all the way in front of cabin

= 0.5 all the way in middle of cabin

= 1.0 all the way in back of cabin

iengloc engine location index

itrim = 0 no changes

= 1 adjust CLh (horizontal tail cl)

= 2 adjust Sh (horizontal tail area)

= 3 adjust xwbox (wing box location)

Outputs:

para[iaxCG] center of gravity

para[iaxCP] center of pressure ( = xCG if itrim=1,2,3 )

para[iaxNP] neutral point location

Future Changes

In an upcoming revision, an aircraft struct and auxiliary indices will be passed in lieu of pre-sliced par arrays.

TASOPT.cglpay — Function

Calculates min and max xCG locations from payload extremes, and corresponding payload fractions.

rfuelF,rpayF give most-forward location xcgF rfuelB,rpayB give most-rearward location xcgB

This version always returns rfuelF = rfuelB = 0.0 which gives an explicit solution for rpayF,rpayB.

The alternative 2D search for rfuel,rpay is kinda ugly, and unwarranted in practice.

TASOPT.htsize — Function

Sets horizontal tail area and wing position to simultaneously:

Meet pitch trim requirement with forward CG

Meet stability requirement with aft CG

Calculates resulting CG, CP, NP locations

Inputs:

pari[.] integer fla array

parg[.] geometry parameter array

paraF[.] aero parameter array for fwdCG case

paraB[.] aero parameter array for aft CG case

paraC[.] aero parameter array for cruise tail CL case

Outputs:

parg[igSh] HT area

parg[igxwbox] wingbox location

parg[igxwing] wing centroid location