Design and evaluation

Sizing the aircraft

The aircraft is sized via a fixed point iteration for the design mission (wsize). The performance of the design can be evaluated for an off-design mission (woper).

wsize is typically the driving script in an analysis, as is the case in the size_aircraft! call (as demonstrated in the first example). The sizing analysis calls the various performance routines (e.g., fusebl, surfw, cdsum, mission, etc.) as shown in the TASOPT flowchart.

TASOPT.wsize — Method

wsize(ac; itermax=35,
wrlx1=0.5, wrlx2=0.9, wrlx3=0.5, initwgt=false, initeng=0, 
iairf=1, Ldebug=false, printiter=true, saveODperf=false)

Main weight sizing function. Calls on various sub-functions to calculate weight of fuselage, wings, tails, etc., and iterates until the MTOW converges to within a specified tolerance.

🔃 Inputs and Outputs

Inputs:

Array of flags that control design choices - fuel types, where to store fuel, etc.

Geometric and structural parameters - dimensions primarily

Aerodynamic parameters - CL, CD, KE dissipation, etc.

Mission-specific parameters - alt, mach, P, T etc.

Engine-specific parameters

Outputs:

No explicit outputs. Computed quantities are saved to par arrays of aircraft model.

TASOPT.woper — Method

woper(ac, mi, itermax, initeng, saveOffDesign)

woper runs the aircraft through input off-design missions

🔃 Inputs and Outputs

Inputs:

ac::aircraft: Aircraft with first mission being the design mission

mi::Int4: Off design mission to run (Default: 1)

itermax::Int64: Maximum iterations for sizing loop

initeng::Boolean: Use design case as initial guess for engine state if true

saveOffDesign::Boolean: Set true if you want computed quanties to be saved in the selected off design par arrays of the aircraft model

Outputs:

No explicit outputs. Computed quantities are saved to par arrays of aircraft model for the off design mission selected

Mission evaluation

A sized aircraft's mission performance can be obtained (mission!), along with operation constraints via a pitch trim calculation (balance) and balanced field length calculation (takeoff!).

TASOPT.mission! — Method

mission!(pari, parg, parm, para, pare, Ldebug)

Runs aircraft through mission, calculating fuel burn and other mission variables.

Input: pari[.] integer flags parg[.] geometry parameters parm[.] mission parameters iairf index of airfoil database to use initeng 0 = engine state will be initialized for all points 1 = engine state is assumed to be initialized ipc1 0 = ipcruise1 aero and engine point needs to be calculated 1 = ipcruise1 aero and engine point assumed calculated

Input/Output: para[.p] aero parameters for points p=1..iptotal pare[.p] engine parameters for points p=1..iptotal

NOTE: This routine assumes that estimates of the climb-leg flight path gamma angles are passed in via para[iagamV,ipclimb1:ipclimbn]. These appear as cos(gamma) factors in the climb equations, and can be passed in as zero with only a minor error. They are updated and returned in the same para[iagamV,ip] array.

!!! compat "Future Changes" In an upcoming revision, an aircraft struct and auxiliary indices will be passed in lieu of pre-sliced par arrays.

TASOPT.takeoff! — Method

  takeoff(ac, initeng, ichoke5, ichoke7)

  Calculates takeoff parameters and balanced field length.
  The aircraft must be defined in parg array. The ipstatic and iprotate points are assumed to exist.

TASOPT.balance — Method

  balance(pari, parg, para, rfuel, rpay, ξpay, itrim)

Makes one of three (or none) changes to achieve pitch trim calculates resulting CG, CP, NP locations.

Inputs:

pari[.] integer flag array

parg[.] geometry parameter array

para[.] aero parameter array

rfuel fuel fraction Wfuelactual/WfuelMTOW

rpay payload fraction Wpayactual/WpayMTOW

ξpay partial-payload packing location

= 0.0 all the way in front of cabin

= 0.5 all the way in middle of cabin

= 1.0 all the way in back of cabin

iengloc engine location index

itrim = 0 no changes

= 1 adjust CLh (horizontal tail cl)

= 2 adjust Sh (horizontal tail area)

= 3 adjust xwbox (wing box location)

Outputs:

para[iaxCG] center of gravity

para[iaxCP] center of pressure ( = xCG if itrim=1,2,3 )

para[iaxNP] neutral point location

Future Changes

In an upcoming revision, an aircraft struct and auxiliary indices will be passed in lieu of pre-sliced par arrays.