Design and evaluation

[Sizing the aircraft] (@id sizing)

The aircraft is sized via a fixed point iteration for the design mission (size_aircraft!()). The performance of the design can be evaluated for the design (imission = 1) and off-design (imission >= 2) missions via fly_mission!().

size_aircraft!() is typically the driving function in an analysis, as in the [first example] (@ref firstexample). The sizing analysis calls the various performance subroutines (e.g., fuselage_drag!(), wing_weights!(), aircraft_drag!(), _mission_iteration!(), etc.) as shown in the TASOPT flowchart. These subroutines are called automatically within _size_aircraft!(), which is wrapped by the user-facing size_aircraft!().

🖥️ Code structure - Aircraft sizing

The aircraft-sizing function requires an aircraft object as input. See read_aircraft_model() to get an idea of the fields that are required in this object. This object is unpacked into storage arrays and other component objects, such as wing, fuselage or engine. The eventual aim is to eliminate all data storage array and replace them by component objects but this is still work in progress.

The first major function called within size_aircraft!() is fuselage_drag!(), which calculates the fuselage boundary layer properties and drag coefficients for start-of-cruise; these are then used in other mission points. size_aircraft!() then uses simplified methods to initialize the relevant aircraft weights and parameters, unless the user specifies otherwise with an optional input (init_weight=true). The bulk of the computational cost and time is spent in the weight sizing loop. After the weight sizing loop is completed, the aircraft takeoff performance and field lengths are calculated using takeoff!().

Weight sizing loop

size_aircraft!() performs a fixed point iteration by sequentially running weight and performance models for the different aircraft components. This is done via a for loop that gets terminated once the maximum aircraft weight has converged within a desired tolerance. The solver will fail to converge for infeasible combinations of aircraft and missions. The solver may fail to converge due to poor initial guesses (or conditioning); this can be addressed by adjusting the initial guess or raising the maximum number of iterations.

The fuselage weight is calculated first in the sizing loop through fusew!(). Then, the total maximum takeoff weight gets recomputed and there is a check for whether the sizing loop is terminated. If weight has not converged, the loop continues.

The wing geometry is set by running set_wing_geometry!() and the wing pitching moments are computed through wing_CM(). The horizontal and vertical tail geometry is computed through tail_loading!(). Finally, the weights of the three aerodynamic surfaces are calculated by running wing_weights!().

If the aircraft requires an insulated fuel tank in the fuselage, for example, if the fuel is cryogenic, the tank is sized using tanksize!(); this function calculates the structural weight and sizes the thermal insulation. For details on how the fuel tank is sized, see Fuel tanks. The sized tank dimensions are then use to recalculate the fuselage geometry to accommodate the tank in update_fuse!().

The weight and balance of the aircraft at start-of-cruise is adjusted using balance_aircraft!(). This can move the wing, resize the horizontal tail, or change the tail trim to achieve a desired metric for longitudinal stability (e.g., a set static margin).

The total drag at start-of-cruise, which is the engine design point, is calculated using aircraft_drag!(). This function calls a combination of models for the drag of aerodynamic surfaces, engine nacelle, and induced drag at the Trefftz plane.

The engines are sized at the start-of-cruise to produce a total thrust force equal to the aircraft drag, as computed by aircraft.engine.enginecalc!(), a specifiable function in the engine object. This field stores a user defined function for the engine performance. Although the user is free to use alternative models by modifying the engine object, TASOPT currently includes a two-spool turbofan engine model. The turbofan engine functions are called via a wrapper, tfwrap!(), which in turns calls the engine calculation function tfcalc!().

Once the engines are sized, the fuel demand at every point in the mission is calculated using _mission_iteration!(). This function in turn recalculates the balance, drag, and engine performance at every point. Further details on mission are provided below. _mission_iteration!() is usually the greatest time sink in an aircraft sizing. The weight gets updated after running _mission_iteration!() and the loop restarts.

[Mission evaluation] (@id mission)

The function _mission_iteration!() contains the fuel burn calculation for the entire mission. It can be used both in sizing, as part of the iteration to obtain a converged aircraft, or in off-design, to calculate the performance of an already-designed airplane. Rather than calling _mission_iteration!() directly, users should call its user-facing wrapper, fly_mission!().

🖥️ Code structure - Mission

The _mission_iteration!() function simulates the entire mission of an aircraft, calculating fuel burn and other mission variables.

From the altitude, the function sets the initial conditions including temperature, pressure, and density atmos(). Then, the lift coefficient is interpolated over the climb and descent points to ensure smooth transitions between different phases of the mission.

Next, the function estimates the takeoff speed and sets the velocity and Reynolds number over the climb and descent points. This involves calculating the takeoff lift coefficient and using it to estimate the takeoff speed. The Reynolds number is then calculated based on the takeoff speed and other atmospheric conditions.

The function proceeds to calculate the climb and descent parameters using aerodynamic and engine performance models. It integrates the climb and descent trajectories using a predictor-corrector scheme to update the range, time, and weight fractions.

Once the climb and descent parameters are set, the function sets the conditions for the cruise phase, including altitude, speed, and fuel consumption. It calculates the fuel burn and weight fractions for the entire mission via calls to engine.enginecalc!(), and adds any vented fuel. This involves adjusting the aircraft's balance and trim settings via calls to balance_aircraft!() to ensure stability throughout the mission, and recalculating the drag via aircraft_drag!().

The function also sets up the climb points at equal altitude intervals and calculates the available thrust assuming maximum-throttle climb. It initializes the climb integrands and integrates the trajectory over the climb phase. The function calculates the cruise-climb angle based on available thrust and atmospheric conditions.

The descent phase is then set up, with the function interpolating the descent points and integrating the time and weight over the descent. It calculates the velocity, Mach number, and Reynolds number for each descent point and adjusts the pitch trim by adjusting the horizontal tail lift coefficient.

Finally, the function calculates the mission fuel fractions and weights, including the weight of any vented fuel; it updates the mission parameters, such as the takeoff and fuel weights, and computes the mission's payload-fuel energy intensity (PFEI; a productivity-specific energy metric in kJ/kg-km),

\[\mathrm{PFEI} = \frac{W_{f,b}\mathrm{LHV}}{g W_{pay} R},\]

where $W_{f,b}$ is the mission fuel burn weight, $\mathrm{LHV}$ is the fuel's lower heating value, $g$ is the acceleration of gravity, $W_{pay}$ is the payload weight, and $R$ is the mission range.

[On- and off-design performance] (@id missionexec)

The function fly_mission!() can calculate the off-design performance of a sized aircraft: it runs the aircraft through a mission with different range, payload, and conditions to the design mission. For this purpose, it calls _mission_iteration!() and iterates the fuel burn until a converged takeoff mass is reached. If called on the design mission, the on-design performance is recovered. See PayloadRange() for an example of how fly_mission!() can be used.

Function documentation

TASOPT.size_aircraft! — Function

size_aircraft(ac::aircraft; iter=35, initwgt=false, Ldebug=false,
    printiter=true, saveOD=false)

sizes the given aircraft instance. A light wrapper around the _size_aircraft! function, which does the actual work.

TASOPT._size_aircraft! — Function

_size_aircraft!(ac; itermax=35,
wrlx1=0.5, wrlx2=0.9, wrlx3=0.5, initwgt=false, initializes_engine=true, 
iairf=1, Ldebug=false, printiter=true)

Main weight sizing function. Calls on various sub-functions to calculate weight of fuselage, wings, tails, etc., and iterates until the MTOW converges to within a specified tolerance. Formerly, wsize().

Warning

_size_aircraft!() Should not be called directly by users, instead use size_aircraft!().

🔃 Inputs and Outputs

Inputs:

Array of flags that control design choices - fuel types, where to store fuel, etc.

Geometric and structural parameters - dimensions primarily

Aerodynamic parameters - CL, CD, KE dissipation, etc.

Mission-specific parameters - alt, mach, P, T etc.

Engine-specific parameters

Outputs:

No explicit outputs. Computed quantities are saved to par arrays of aircraft model.

TASOPT.fly_mission! — Function

fly_mission!(ac, imission, itermax, initializes_engine, opt_prescribed_cruise_parameter)

Runs the aircraft through the specified mission, computing and converging the fuel weight. Formerly, fly_offdesign_mission!().

🔃 Inputs and Outputs

Inputs:

ac::aircraft: Aircraft with first mission being the design mission

imission::Int64: Off design mission to run (Default: 1)

itermax::Int64: Maximum iterations for sizing loop

initializes_engine::Boolean: Use design case as initial guess for engine state if true

opt_prescribed_cruise_parameter::String: option for whether cruise altitude or lift coefficient is specified. Options are "altitude" or "lift_coefficient"

Outputs:

No explicit outputs. Computed quantities are saved to par arrays of aircraft model for the mission selected

TASOPT.aerodynamics.fuselage_drag! — Function

fuselage_drag!(fuse, parg, para, ip)

Calculates surface velocities, boundary layer, wake for a quasi-axisymmetric body in compressible flow. Formerly, fusebl!().

A compressible source line represents the potential flow. An integral BL formulation with lateral divergence represents the surface BL and wake. An added-source distribution represents the viscous displacement influence on the potential flow. The body shape is defined by its area and perimeter distributions A(x), b0(x), which are defined by the various geometric parameters in parg.

🔃 Inputs and Outputs

Inputs:

fuse::Fuselage: Fuselage model object.

parg::AbstractArray{Float64}: Vector of aircraft model geometry parameters.

para::AbstractArray{Float64}: Vector of aircraft model aerodynamic parameters.

parm::AbstractArray{Float64}: Vector of aircraft model mission parameters.

ip::Integer: Index of flight point in par arrays.

Outputs:

No explicit outputs. Computed drag values are saved to para of aircraft model.

See Simplified Viscous/Inviscid Analysis for Nearly-Axisymmetric Bodies. See also _axisymm_BL and _axisymm_flow.

Future Changes

In an upcoming revision, an aircraft struct and auxiliary indices will be passed in lieu of pre-sliced par arrays.

TASOPT.aerodynamics.set_wing_geometry! — Function

set_wing_geometry!(W, CL, qinf, wing)

Sizes wing area, span, root chord from q, CL, W, AR at given point (taken as start-of-cruise in size_aircraft!).

🔃 Inputs and Outputs

Inputs:

W::Float64: Aircraft weight.

CL::Float64: Lift coefficient.

qinf::Float64: Freestream dynamic head.

wing::TASOPT.structures.Wing: Wing structure

See Sections 2.5 and 3.4.1 of the TASOPT Technical Desc.

TASOPT.aerodynamics.wing_CM — Function

wing_CM(b,bs,bo, sweep, Xaxis,
                   λt, λs, γt, γs,
                   AR, fLo, fLt, cmpo, cmps, cmpt)

Calculates components of wing pitching moment ($C_M$) about wing root axis:

$C_M = C_{M,0} + C_{M,1} (C_L - C_{L,surf})$

$ΔC_{m, surf} = ΔC_{m, 0} + dCₘ/dCL × (C_L - C_{L,h})$

Formerly, surfcm().

🔃 Inputs and Outputs

Inputs:

b::Float64: Span.

bs::Float64: Outer panel break span.

bo::Float64: Root (fuselage) span.

sweep::Float64: Sweep, degrees.

- Xaxis::Float64: Surface axis position.

λt::Float64: Outer-panel chord taper ratio ct/co.

λs::Float64: Inner-panel chord taper ratio cs/co.

γt::Float64: Outer-panel load taper ratio pt/po.

γs::Float64: Inner-panel load taper ratio ps/po.

AR::Float64: Surface aspect ratio.

fLo::Float64, fLt::Float64 : Wing root and tip load adjustment factors.

cmpo::Float64,cmps::Float64,cmpt::Float64: Perpendicular sectional lift coefficient at wing root, break ("snag"), and tip.

Outputs:

CM0::Float64: Zero-lift surface pitching moment.

CM1::Float64: Surface pitching moment including lift contribution.

See Section 2.6.3 of the TASOPT Technical Desc. See also wing_profiledrag_scaled and wing_profiledrag_direct.

TASOPT.aerodynamics.tail_loading! — Function

tail_loading!(tail,S,qne; t_fac = 1.0)

Calculates stabilizer span, root chord, and root loading based on the never-exceed dynamic pressure, maximum CL, sweep, and aspect ratio. Formerly, tailpo!().

🔃 Inputs and Outputs

Inputs:

tail::TASOPT.structures.tail: Tail structure.

S::Float64: Stabilizer area.

qne::Float64: Never-exceed dynamic pressure.

t_fac::Float64: Tail Factor (1 for Htail/Wing, 2 for Vtail).

Outputs:

po::Float64: Stabilizer root loading.

b::Float64: Stabilizer wingspan.

See Geometry or Section 2.3.2 and 2.9.6 of the TASOPT Technical Description.

TASOPT.structures.wing_weights! — Function

wing_weights!(wing, po, gammat, gammas, 
   Nload, We, neout, dyeout, neinn, dyeinn, sigfac, rhofuel; n_wings=2.0)

Calculates Wing or Tail loads, stresses, weights of individual wing sections. Also returns the material gauges, torsional and bending stiffness. Formerly, get_wing_weights!().

🔃 Inputs and Outputs

Inputs:

wing::TASOPT.structures.Wing: Wing structure.

po::Float64: Wing root loading magnitude.

gammat::Float64: Wing outer section taper ratio.

gammas::Float64: Wing inner section taper ratio.

Nload::Int: Max vertical load factor for wing bending loads.

We::Float64: Weight of the engine.

neout::Int: Number of engines mounted outboard of the wing break (strut attachment point).

dyeout::Float64: Spanwise moment arm for outboard engines, measured from the wing break station.

neinn::Int: Number of engines mounted inboard of the wing break (strut attachment point).

dyeinn::Float64: Spanwise moment arm for inboard engines, measured from the wing break station.

sigfac::Float64: Stress Factor.

rhofuel::Float64: Density of the fuel.

n_wings::Int64: Number of total wings (1 for Vtail).

See Geometry, Wing/Tail Structures, and Section 2.7 of the TASOPT Technical Description.

TASOPT.structures.update_fuse! — Function

update_fuse!(ac, imission)

Function to update the fuselage layout when there is a change in fuselage fuel-tank length.

🔃 Inputs and Outputs

Inputs:

ac::aircraft: object with aircraft parameters

imission::Int64: mission index

Outputs: No direct outputs; parameters in ac are modified.

TASOPT.engine.tfwrap! — Function

tfwrap!(ac, case, imission, ip, initializes_engine, iterw = 0)

Calls the turbofan sizing or off-design performance functions for the aircraft's turbofan model. This function is basically a wrapper on tfcalc!, going from the basic engine inputs to those required by the function and storing the outputs.

🔃 Inputs and Outputs

Input:

ac::aircraft: aircraft object

case::String: case identifier, e.g. "sizing" or "off_design"

imission::Int64: mission index

ip::Int64: mission point index

initializes_engine::Bool: flag to initialize engine:

true: initialize variables for iteration in engine

false: use current variables as initial guesses in engine

iterw::Int64: sizing loop iteration

Output: No direct outputs. The ac object gets modified with the engine parameters.

TASOPT._mission_iteration! — Method

_mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)

Runs aircraft through mission, calculating fuel burn and other mission variables. Formerly, mission!().

🔃 Inputs and Outputs

Inputs:

ac::aircraft: aircraft data storage object

imission::Int64: mission index

Ldebug::Bool: debugging flag

NOTE: This routine assumes that estimates of the climb-leg flight path gamma angles are passed in via para[iagamV,ipclimb1:ipclimbn]. These appear as cos(gamma) factors in the climb equations, and can be passed in as zero with only a minor error. They are updated and returned in the same para[iagamV,ip] array.

TASOPT.takeoff! — Method

takeoff!(ac)

Calculates takeoff parameters and balanced field length. The aircraft must be defined in parg array. The ipstatic and iprotate points are assumed to exist.

TASOPT.PayloadRange — Function

PayloadRange(ac_og; Rpts, Ppts, plots_OEW, filename, itermax, initializes_engine, Ldebug)

Function to plot a payload range diagram for an aircraft

🔃 Inputs and Outputs

Inputs:

ac_og::aircraft: Aircraft structure for payload range diagram.

Rpts::Int64: Density of ranges to be plot (Optional).

Ppts::Int64: Density of payloads to be plot (Optional).

filename::String: filename string for the plot to be stored (Optional).

plots_OEW::Bool: Whether to plot OEW+Payload (true) on the y-axis or just Payload (false, default) (Optional).

itermax::Int64: Max Iterations for fly_mission! loop (Optional).

initializes_engine::Bool: Use design case as initial guess for engine state if true (Optional)

specifying_cruise::String: option for whether cruise altitude or lift coefficient is specified. Options are "altitude" or "lift_coefficient"

Ldebug::Bool: verbosity flag. false by default, hiding outputs as PR sweeps progress (Optional).